Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main(){
-    int a,b,c;
+    float a,b,c;
     float Average;
-    scanf("%d %d %d %f", &a, &b, &c, &Average);
+    scanf("%f %f %f %f", &a, &b, &c, &Average);
     Average=(a+b+c)/3;
     printf("Average: %.2f",Average);
     return 0;
